Change all time vars to uint64_t

+++ control/bdot/bdot_control.h
@@ -17,6 +17,8 @@
 
 #include "adcs_math/vector.h"
 
+#include <stdint.h>
+
 
 /**@brief Calculate a dipole command for all three coils given our
  *  magnetic field and angular velocity data.
@@ -34,7 +36,7 @@
 void bdot_control(
     vec3    mag,
     vec3    last_mag,
-    double  delta_t,
+    uint64_t  delta_t,
     vec3   *coils_current
 );
